Moves tileMap.js
